requirements.txt
pygad>=3.2.0
PyYAML>=6.0.1

